Revert "Update tests"
This reverts commit 2ded026e0733b293ef7434932cb2080b9242d362.
