+++ test/conftest.py
@@ -50,9 +50,9 @@ def mock_callback_data(monkeypatch, tmp_path):
     filename = Path(tmp_path) / "temporary.json"
 
     def _mock_dumps(data, *a, **kw):
-        if isinstance(data, dict) and "files" in data:
+        with open(filename, "w") as outfile:
             # We are only going to dump data into temporary file, that we 'know' are generated by us.
-            with open(filename, "w") as outfile:
+            if isinstance(data, dict) and "files" in data:
                 json_data = real_dumps(data)
                 outfile.write(json_data)
         return real_dumps(data, *a, **kw)
@@ -83,8 +83,8 @@ def mock_reuse_zarr(monkeypatch):
     monkeypatch.setattr(ng, "bioformats_gen_zarr", _mock_bioformats_gen_zarr)
 
 
-@pytest.fixture(autouse=True, scope="session")
-def prefect_test_fixture():
+@pytest.fixture
+def prefect_test():
     """
     Reference: https://docs.prefect.io/2.13.5/guides/testing/
     """
