+++ test/input_files/.gitattributes
@@ -5,4 +5,3 @@
 *jpeg filter=lfs diff=lfs merge=lfs -text
 *jpg filter=lfs diff=lfs merge=lfs -text
 *.mrc filter=lfs diff=lfs merge=lfs -text
-*.czi filter=lfs diff=lfs merge=lfs -text
